windows support: move timeBeginPeriod(), remove timeEndPeriod()
There's not the slightest reason why timeBeginPeriod() should not be in
InitTimer(), which was made to hide away OS-specific timer calls. Make
sure it's called after SetErrorMode() just to be sure, though.

Remove the timeEndPeriod() call, which is kind of an eye-sore. According
to MSDN, Begin/End calls must be matched, but I since this is at program
termination anyway, I don't care.